refactor: move frontend utils to test config
